requirements.txt
annotated-types==0.6.0
blinker==1.7.0
cachelib==0.10.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.0
cycler==0.12.1
Flask==3.0.0
Flask-Plots==0.0.1
Flask-Session==0.5.0
fonttools==4.47.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.2
mpld3==0.5.9
numpy==1.26.2
packaging==23.2
pandas==2.1.4
pandas-stubs==2.1.4.231218
Pillow==10.1.0
pydantic==2.5.2
pydantic-settings==2.1.0
pydantic_core==2.14.5
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
redis==5.0.1
requests==2.31.0
scikit-learn==1.3.2
scipy==1.11.4
seaborn==0.13.0
six==1.16.0
spotipy==2.23.0
threadpoolctl==3.2.0
types-pytz==2023.3.1.1
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.1.0
Werkzeug==3.0.1
